+++ src/features/Treege/type/TreeNode.ts
@@ -9,9 +9,24 @@ export interface TreeValues {
   message?: string;
 }
 
+export interface Params {
+  id: string;
+  key: string;
+  value: string;
+}
+
+export interface PathKey {
+  object?: string;
+  value?: string;
+  label?: string;
+  image?: string;
+}
+
 export interface Route {
   url: string;
   searchKey: string;
+  pathKey?: PathKey;
+  params?: Params[];
 }
 
 export interface TreeNode extends Omit<RawNodeDatum, "attributes" | "children"> {