+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/treege/releases)
 
-## v0.13.1
-- **[fix]** : Error semantic html
+## v0.14.0
+- **[feat]** : Add autocomplete pathKey for object schema
+- **[feat]** : Add possibility to add params to autocomplete